refactor tests for pystiche.pyramid (#374)
